+++ test/exactly_lib_test/test_case_utils/file_matcher/name/reg_ex.py
@@ -1,13 +1,11 @@
 import pathlib
 import unittest
 
-from exactly_lib.symbol.symbol_syntax import symbol_reference_syntax_for_name
 from exactly_lib.util.logic_types import ExpectationType
-from exactly_lib_test.section_document.test_resources.parse_source import remaining_source
 from exactly_lib_test.symbol.test_resources.string import StringSymbolContext
 from exactly_lib_test.symbol.test_resources.symbols_setup import SymbolContext
 from exactly_lib_test.test_case_utils.file_matcher.test_resources import argument_building as arg
-from exactly_lib_test.test_case_utils.file_matcher.test_resources import argument_syntax, integration_check
+from exactly_lib_test.test_case_utils.file_matcher.test_resources import integration_check
 from exactly_lib_test.test_case_utils.file_matcher.test_resources import parse_test_base_classes as test_case_utils
 from exactly_lib_test.test_case_utils.file_matcher.test_resources.integration_check import ARBITRARY_MODEL
 from exactly_lib_test.test_case_utils.file_matcher.test_resources.test_utils import Actual
@@ -15,84 +13,73 @@ from exactly_lib_test.test_case_utils.logic.test_resources.intgr_arr_exp import 
     ExecutionExpectation, Expectation
 from exactly_lib_test.test_case_utils.parse.test_resources.arguments_building import Arguments
 from exactly_lib_test.test_case_utils.regex.parse_regex import is_reference_to_valid_regex_string_part
-from exactly_lib_test.test_case_utils.regex.test_resources.argument_syntax import reg_ex_args_list
 from exactly_lib_test.test_case_utils.regex.test_resources.validation_cases import failing_regex_validation_cases
-from exactly_lib_test.test_case_utils.test_resources.negation_argument_handling import \
-    ExpectationTypeConfigForNoneIsSuccess, PassOrFail
 from exactly_lib_test.test_resources.matcher_argument import NameRegexComponent
 from exactly_lib_test.test_resources.test_utils import NEA
 from exactly_lib_test.test_resources.value_assertions import value_assertion as asrt
+from exactly_lib_test.type_system.trace.test_resources import matching_result_assertions as asrt_matching_result
 
 
 def suite() -> unittest.TestSuite:
     return unittest.TestSuite([
-        TestRegExPatternOnBaseName(),
+        unittest.makeSuite(TestRegExPatternOnBaseName),
         ParseShouldFailWhenRegexArgumentIsMissing(),
         TestWithSymbolReferences(),
         ValidationShouldFailWhenRegexIsInvalid(),
     ])
 
 
-def name_matches_regex_arg(regex: str) -> arg.FileMatcherArg:
-    return arg.Name(arg.NameRegexVariant(NameRegexComponent(regex)))
-
-
 class TestRegExPatternOnBaseName(test_case_utils.TestCaseBase):
-    def runTest(self):
+    def test_base_name(self):
         # ARRANGE #
         cases = [
-            NEA('match basename with exact match',
+            NEA('exact match',
                 True,
                 Actual(
-                    Arguments(argument_syntax.base_name_reg_ex_pattern_matcher_of(
-                        reg_ex_args_list('.*name'),
-                    )),
-                    pathlib.Path('file name'),
+                    name_matches_regex_args('name'),
+                    pathlib.Path('name'),
                 ),
                 ),
-            NEA('match basename with substring match',
+            NEA('exact match, with ignored dir component',
                 True,
                 Actual(
-                    Arguments(argument_syntax.base_name_reg_ex_pattern_matcher_of(
-                        reg_ex_args_list('PA..ERN')
-                    )),
-                    pathlib.Path('before PATTERN after'),
+                    name_matches_regex_args('base-name'),
+                    pathlib.Path('dir-name') / 'base-name',
                 ),
                 ),
-            NEA('match name with pattern that matches path components',
+            NEA('exact non-match',
                 False,
                 Actual(
-                    Arguments(argument_syntax.base_name_reg_ex_pattern_matcher_of(
-                        reg_ex_args_list('dir-name')
-                    )),
-                    pathlib.Path('dir-name') / pathlib.Path('base-name'),
+                    name_matches_regex_args('expected'),
+                    pathlib.Path('dir-name') / 'actual',
                 ),
                 ),
-            NEA('not match, because pattern is not in path',
-                False,
+            NEA('pattern 1',
+                True,
                 Actual(
-                    Arguments(argument_syntax.base_name_reg_ex_pattern_matcher_of(
-                        reg_ex_args_list('PATTERN')
-                    )),
-                    pathlib.Path('not the pattern'),
+                    name_matches_regex_args('.*name$'),
+                    pathlib.Path('dir-name') / 'file name',
+                ),
+                ),
+            NEA('pattern 2',
+                True,
+                Actual(
+                    name_matches_regex_args('PA..ERN'),
+                    pathlib.Path('dir-name') / 'before PATTERN after',
                 ),
                 ),
             NEA('not match, because pattern with unexpected case is in base name',
                 False,
                 Actual(
-                    Arguments(argument_syntax.base_name_reg_ex_pattern_matcher_of(
-                        reg_ex_args_list('base name')
-                    )),
-                    pathlib.Path('THE BASE NAME'),
+                    name_matches_regex_args('base name'),
+                    pathlib.Path('BASE NAME'),
                 ),
                 ),
             NEA('match, because pattern with unexpected case is in base name, but test is case insensitive',
                 True,
                 Actual(
-                    Arguments(argument_syntax.base_name_reg_ex_pattern_matcher_of(
-                        reg_ex_args_list('base name',
-                                         ignore_case=True)
-                    )),
+                    name_matches_regex_args('base name',
+                                            ignore_case=True),
                     pathlib.Path('THE BASE NAME'),
                 ),
                 ),
@@ -102,30 +89,50 @@ class TestRegExPatternOnBaseName(test_case_utils.TestCaseBase):
 
         self._check_cases__with_source_variants(cases)
 
+    def test_fail_when_pattern_matches_dir_components(self):
+        # ARRANGE #
+        cases = [
+            NEA('non-match - name with pattern that matches path components',
+                False,
+                Actual(
+                    name_matches_regex_args('dir-name'),
+                    pathlib.Path('dir-name') / pathlib.Path('base-name'),
+                ),
+                ),
+            NEA('non-match - name with pattern that matches path components 2',
+                False,
+                Actual(
+                    name_matches_regex_args('name.*name'),
+                    pathlib.Path('dir-name') / pathlib.Path('base-name'),
+                ),
+                ),
+        ]
 
-class ParseShouldFailWhenRegexArgumentIsMissing(test_case_utils.TestWithNegationArgumentBase):
-    argument_w_opt_neg = arg.WithOptionalNegation(
-        arg.Custom('')
-    )
+        # ACT & ASSERT #
 
-    def _doTest(self, maybe_not: ExpectationTypeConfigForNoneIsSuccess):
+        self._check_cases__with_source_variants(cases)
+
+
+class ParseShouldFailWhenRegexArgumentIsMissing(test_case_utils.TestCaseBase):
+    def runTest(self):
+        arguments = arg.Name(arg.NameRegexVariant(
+            NameRegexComponent('')
+        ))
         self._assert_failing_parse(
-            remaining_source(str(self.argument_w_opt_neg.get(maybe_not.expectation_type)))
+            arguments.as_remaining_source
         )
 
 
 class ValidationShouldFailWhenRegexIsInvalid(test_case_utils.TestCaseBase):
     def runTest(self):
         for regex_case in failing_regex_validation_cases():
-            argument_w_opt_neg = arg.WithOptionalNegation(
-                name_matches_regex_arg(regex_case.regex_string)
-            )
+            arguments = name_matches_regex_args(regex_case.regex_string)
             for expectation_type in ExpectationType:
                 with self.subTest(expectation_type=expectation_type,
                                   validation_case=regex_case.case_name):
                     self._check_with_source_variants(
                         arguments=
-                        Arguments(str(argument_w_opt_neg.get(expectation_type))),
+                        arguments,
                         model_constructor=
                         ARBITRARY_MODEL,
                         arrangement=
@@ -144,32 +151,37 @@ class ValidationShouldFailWhenRegexIsInvalid(test_case_utils.TestCaseBase):
                     )
 
 
-class TestWithSymbolReferences(test_case_utils.TestWithNegationArgumentBase):
-    any_char_regex_string_symbol = StringSymbolContext.of_constant(
-        'valid_regex_string_symbol',
-        '.',
-    )
-    argument_w_opt_neg = arg.WithOptionalNegation(
-        name_matches_regex_arg('AB' + symbol_reference_syntax_for_name(any_char_regex_string_symbol.name))
-    )
-    arrangement = arrangement_w_tcds(
-        symbols=any_char_regex_string_symbol.symbol_table)
-
-    def _doTest(self, maybe_not: ExpectationTypeConfigForNoneIsSuccess):
+class TestWithSymbolReferences(test_case_utils.TestCaseBase):
+    def runTest(self):
+        any_char_regex_string_symbol = StringSymbolContext.of_constant(
+            'valid_regex_string_symbol',
+            '.',
+        )
+        arguments = name_matches_regex_args('AB' + any_char_regex_string_symbol.name__sym_ref_syntax)
         self._check_with_source_variants(
             arguments=
-            Arguments(str(self.argument_w_opt_neg.get(maybe_not.expectation_type))),
+            arguments,
             model_constructor=
             integration_check.constant_relative_file_name('ABC'),
-            arrangement=self.arrangement,
+            arrangement=arrangement_w_tcds(
+                symbols=any_char_regex_string_symbol.symbol_table
+            ),
             expectation=Expectation(
                 ParseExpectation(
                     symbol_references=asrt.matches_sequence([
-                        is_reference_to_valid_regex_string_part(self.any_char_regex_string_symbol.name),
+                        is_reference_to_valid_regex_string_part(any_char_regex_string_symbol.name),
                     ]),
                 ),
                 ExecutionExpectation(
-                    main_result=maybe_not.main_result(PassOrFail.PASS)
+                    main_result=asrt_matching_result.matches_value(True)
                 ),
             )
         )
+
+
+def name_matches_regex_args(regex: str,
+                            ignore_case: bool = False,
+                            ) -> Arguments:
+    return arg.Name(
+        arg.NameRegexVariant(NameRegexComponent(regex, ignore_case))
+    ).as_arguments