+++ test/exactly_lib_test/test_case_utils/file_matcher/z_package_suite.py
@@ -1,16 +1,16 @@
 import unittest
 
-from exactly_lib_test.test_case_utils.file_matcher import name_glob_pattern, name_reg_ex, file_type, std_expr, \
+from exactly_lib_test.test_case_utils.file_matcher import file_type, std_expr, \
     contents_of_file
 from exactly_lib_test.test_case_utils.file_matcher.contents_of_dir import z_package_suite as contents_of_dir
+from exactly_lib_test.test_case_utils.file_matcher.name import z_package_suite as name
 from exactly_lib_test.test_case_utils.file_matcher.run_program import z_package_suite as run_program
 
 
 def suite() -> unittest.TestSuite:
     return unittest.TestSuite([
         std_expr.suite(),
-        name_glob_pattern.suite(),
-        name_reg_ex.suite(),
+        name.suite(),
         file_type.suite(),
         contents_of_file.suite(),
         contents_of_dir.suite(),